campuscashcontrolb/requirements.txt
asgiref==3.11.1
Django==5.1.7
django-cors-headers==4.9.0
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
et_xmlfile==2.0.0
numpy==2.4.4
openpyxl==3.1.5
packaging==26.0
pandas==3.0.2
psycopg2-binary==2.9.11
PyJWT==2.12.1
python-dateutil==2.9.0.post0
setuptools==82.0.1
six==1.17.0
sqlparse==0.5.5
tzdata==2026.1
wheel==0.46.3
xlrd==2.0.2
